Small refactor, so that ILP BN can be implemented
