+++ models/DynamicBayesianNetwork.py
@@ -1,12 +1,6 @@
 from abc import ABC, abstractmethod
-from pgmpy.models import BayesianModel
-from pgmpy.estimators import BayesianEstimator
 from PGM import PGM
-from pgmpy.inference import VariableElimination
-from pgmpy.estimators import HillClimbSearch
-from pgmpy.estimators import BicScore
-from pgmpy.sampling import GibbsSampling
-
+import pandas as pd
 
 class DynamicBayesianNetwork(PGM, ABC):
     """Abstract class for dynamic Bayesian networks (DBNs). Generally, DBNs consist of initial Bayesian network which
@@ -20,3 +14,4 @@ class DynamicBayesianNetwork(PGM, ABC):
         structure_and_weights (bool): Indicates that the model is capable of learning the structure and weights together.
         """
         super().__init__(structure_and_weights)
+
